backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
psycopg2-binary==2.9.11
pillow==12.0.0
python-dotenv==1.1.1
flake8==7.3.0